refactor(runtime): split build_client_from_entry into llm/client_builder + fix guarded_http error fidelity
Addresses two Task 7 review findings: supervisor_runner.rs exceeded the 800-line
mandate (moved client construction to llm/client_builder.rs), and a guarded_http
build failure for local/ollama was mislabeled "unsupported provider" (restored
pre-extraction ordering so the real error propagates).
