+++ mur-agent-runtime/src/llm/mod.rs
@@ -4,6 +4,7 @@ use async_trait::async_trait;
 use mur_common::{AgentProfile, LlmMode};
 
 pub mod anthropic;
+pub(crate) mod client_builder;
 pub mod fallback;
 pub mod ollama;
 pub mod openai;
